Fixed major problems with auto-evo tolerance changes (#6421)
* Fixed major problems with auto-evo tolerance changes

mainly it not working at all and nonsensical adjustments that it would have calculated. Added also safety checking code and switched tolerance scores to be doubles to avoid imprecision problems from slightly off stats

* Added a warning print

+++ src/auto-evo/selection_pressure/EnvironmentalTolerancePressure.cs
@@ -39,11 +39,13 @@ public class EnvironmentalTolerancePressure : SelectionPressure
 
     public override float Score(Species species, Patch patch, SimulationCache cache)
     {
+        // TODO: multicellular tolerance
         if (species is not MicrobeSpecies microbeSpecies)
             return 0;
 
         // Use scores to encourage species to be adapted to their environment
-        return MicrobeEnvironmentalToleranceCalculations.CalculateTotalToleranceScore(microbeSpecies, patch.Biome);
+        return (float)MicrobeEnvironmentalToleranceCalculations.CalculateTotalToleranceScore(microbeSpecies,
+            patch.Biome);
     }
 
     public override float GetEnergy(Patch patch)
